Refactor API routes and imports, remove unused icons
