+++ pet-management-app/src/app/api/admin/settings/route.ts
@@ -11,7 +11,7 @@ interface SystemSetting {
   category: string
 }
 
-export async function GET(request: NextRequest) {
+export async function GET() {
   try {
     const session = await getServerSession(authOptions)
