+++ pet-management-app/src/app/api/ai-vet/status/route.ts
@@ -15,7 +15,7 @@ async function getSystemSetting(key: string, defaultValue: string): Promise<stri
   }
 }
 
-export async function GET(request: NextRequest) {
+export async function GET() {
   try {
     const session = await getServerSession(authOptions)
