+++ pet-management-app/src/app/api/reminders/route.ts
@@ -13,7 +13,7 @@ export async function GET(request: NextRequest) {
     const { searchParams } = new URL(request.url)
     const status = searchParams.get('status') // 'active', 'completed', 'all'
 
-    let where: any = {
+    const where: Record<string, unknown> = {
       pet: {
         userId: session.user.id
       }
